requirements.txt
python-qpid-proton
configparser
minio